requirements/prod.txt
click-option-group==0.5.3
click==7.1.2
dataclasses
dataclasses-json==0.5.5
face-recognition
ffmpeg-python
h5py==2.10.0
inquirer==2.7.0
librosa==0.8.1
matplotlib
more-itertools==8.7.0
mypy==0.812
numpy==1.16.4  # this exact version is needed to make tensorflow/gpu happy!
opencv-python
pandas==1.1.5
Pillow==6.2.1
pooch==1.5.1  # To prevent warnings from scikit-image
pyocclient==0.6
requests==2.22.0
resampy==0.2.2
scikit-image
scikit-learn
scipy==1.3.3
tensorflow-gpu==1.14